+++ pueue/tests/daemon/integration/remove.rs
@@ -31,7 +31,14 @@ async fn test_normal_remove() -> Result<()> {
     pause_tasks(shared, TaskSelection::TaskIds(vec![3])).await?;
 
     // Stash task 5
-    send_message(shared, Message::Stash(vec![5])).await?;
+    send_message(
+        shared,
+        StashMessage {
+            tasks: TaskSelection::TaskIds(vec![5]),
+            enqueue_at: None,
+        },
+    )
+    .await?;
 
     let remove_message = Message::Remove(vec![0, 1, 2, 3, 4, 5]);
     send_message(shared, remove_message).await?;